ai-engineer/07-advanced-rag/cohere-reranking/python/requirements.txt
langchain-core
langchain-classic
langchain-huggingface
langchain-chroma
langchain-cohere
sentence-transformers
chromadb
python-dotenv
